Drop 147 redundant casts that restate the destination's own declared type
Extends the call-argument cast sweep to assignment destinations. Where a
read-modify-write assigns back into the field or global it just read --
state->motionFlags = (u8)(state->motionFlags | MASK) -- the cast spells out the
conversion the assignment already performs, so it is decompilation
compensation rather than source. Restricted to destinations whose declared type
is unique repo-wide and equal to the cast type, to expressions with no call and
no address-of (a pointer-to-integer cast is load-bearing), and to fields and
globals only: a cast into a narrow local really does emit a truncation the
plain assignment omits. Casts covering only part of the right-hand side, and
the nested conversions inside these expressions, are untouched.

Removing the cast leaves the outer parentheses wrapping nothing, so they go
too, together with six pre-existing assignments that already had that shape.
Parentheses around a boolean comparison are kept.

Byte-verified per file with automatic revert; nothing regressed, so all 46
edited files are kept. All 991 source objects identical to the pre-change
build, ninja all_source and ninja EXIT=0, dtk shasum main.dol OK, report.json
fuzzy 99.776726 / complete_units 881, both unchanged.

+++ src/dlls/engine/10_expgfx/expgfx.c
@@ -136,7 +136,7 @@ static inline u32 Expgfx_GetSlotTableIndex(const ExpgfxSlot* slot)
 
 static inline void Expgfx_SetSlotTableIndex(ExpgfxSlot* slot, u8 tableIndex)
 {
-    slot->encodedTableIndex = (u8)((tableIndex << 1) | (slot->encodedTableIndex & 1));
+    slot->encodedTableIndex = (tableIndex << 1) | (slot->encodedTableIndex & 1);
 }
 
 static inline ExpgfxSlot* Expgfx_GetSlot(int poolIndex, int slotIndex)
@@ -2354,11 +2354,11 @@ void expgfx_initSlotQuad(void* slotPtr)
 
     if ((slot->behaviorFlags & EXPGFX_BEHAVIOR_SCALE_FROM_ZERO) != 0)
     {
-        slot->scaleCurrent = ((f32)slot->scaleStep * step + (f32)slot->scaleCurrent);
+        slot->scaleCurrent = (f32)slot->scaleStep * step + (f32)slot->scaleCurrent;
     }
     else if ((slot->renderFlags & EXPGFX_RENDER_SCALE_OVER_LIFETIME) != 0)
     {
-        slot->scaleCurrent = ((f32)slot->scaleCurrent - (f32)slot->scaleStep * step);
+        slot->scaleCurrent = (f32)slot->scaleCurrent - (f32)slot->scaleStep * step;
     }
 
     if (resource == 0)
