+++ src/dlls/objects/332/332.c
@@ -519,7 +519,7 @@ void babyCloudRunner_init(GameObject* obj, BabyCloudRunnerPlacement* placement) 
     if (mainGetBit(placement->runnerGameBit) != 0) {
         ObjHits_DisableObject(obj);
         obj->anim.flags = (s16)(obj->anim.flags | OBJANIM_FLAG_HIDDEN);
-        state->captureFlags = (u8)(state->captureFlags & ~BABYCLOUDRUNNER_CAPTURE_ACTIVE);
+        state->captureFlags = state->captureFlags & ~BABYCLOUDRUNNER_CAPTURE_ACTIVE;
         Obj_RemoveFromUpdateList(obj);
         ObjGroup_RemoveObject((int)obj, BABYCLOUDRUNNER_PRIMARY_OBJECT_GROUP);
     } else {
