+++ src/dlls/objects/396_MMSH_Shrine/MMSH_Shrine.c
@@ -168,9 +168,9 @@ void mmshShrine_updateHoverMotion(GameObject* obj) {
         return;
     }
 
-    state->orbitPhaseA = (s16)(state->orbitPhaseA + (int)(MMSH_SHRINE_ORBIT_RATE_A * timeDelta));
-    state->orbitPhaseB = (s16)(state->orbitPhaseB + (int)(MMSH_SHRINE_ORBIT_RATE_B * timeDelta));
-    state->orbitPhaseC = (s16)(state->orbitPhaseC + (int)(MMSH_SHRINE_ORBIT_RATE_C * timeDelta));
+    state->orbitPhaseA = state->orbitPhaseA + (int)(MMSH_SHRINE_ORBIT_RATE_A * timeDelta);
+    state->orbitPhaseB = state->orbitPhaseB + (int)(MMSH_SHRINE_ORBIT_RATE_B * timeDelta);
+    state->orbitPhaseC = state->orbitPhaseC + (int)(MMSH_SHRINE_ORBIT_RATE_C * timeDelta);
 
     obj->anim.localPosY =
         MMSH_SHRINE_ORBIT_HEIGHT +
